Fix NaN loss: align model dtype with the trainer's autocast mode
Upstream hardcodes `self.lang_model.half()` (fp16) while the config passes
--bf16 True, so HF Trainer wraps the forward in a bfloat16 autocast. That worked
upstream only because DeepSpeed reconciled precision (ds_configs/stage2.json sets
"bf16": {"enabled": "auto"}). Running without DeepSpeed leaves fp16 weights under
a bf16 autocast, and the loss is NaN from the first step.

Three precisions were in play on the input-embedding path: the LLM in fp16,
MyEmbedding's vision encoder in fp32, and MyEmbedding.weight tied to the fp16
word embedding.

- Reg2RG picks one dtype (bf16 on Ampere+, fp16 fallback) and applies it to both
  the LLM and MyEmbedding. MyEmbedding is cast before the weight tie: the tie
  shares a Parameter object, so casting afterwards would re-cast the LLM's
  embedding as a side effect.
- my_embedding_layer allocated vision_region_embedding at torch's fp32 default;
  the following torch.cat then promoted the whole embedding table back to fp32
  mid-graph. It now follows self.weight.dtype.
- The collator padded absent regions with fp32 zeros regardless of the real
  tensors' dtype; it now carries the reference dtype.

+++ src/Model/my_embedding_layer.py
@@ -179,10 +179,14 @@ class MyEmbedding(nn.Module):
         for key in region_embeddings.keys():
             region_embeddings[key] = torch.cat([region_embeddings[key], mask_embeddings[key].unsqueeze(1)], dim=1) 
         
-        max_region = len(region_embeddings)  
+        max_region = len(region_embeddings)
         # zero init the vision embedding
+        # NOTE: dtype must follow the module, not torch's fp32 default. Otherwise
+        # this buffer is fp32 while self.weight is bf16, and the torch.cat below
+        # promotes the whole embedding table back to fp32 mid-graph.
         vision_region_embedding = torch.zeros(
-            (B, 33*max_region, self.embedding_dim), device=text_input.device) # NOTE: 2 means 1 region and 1 mask embeddings
+            (B, 33*max_region, self.embedding_dim),
+            device=text_input.device, dtype=self.weight.dtype) # NOTE: 2 means 1 region and 1 mask embeddings
         
         for i in range(B):
             for j in range(len(region2areas[i])):
